fix: Yazio recipe unit parsing inflation + AI recipe decimal-calories crash
- RecipeNutritionAnalyzer: add full-word units (gram, milliliter, tablespoon, teaspoon...) to UNIT_TO_G, which were previously unrecognized and fell through to a '* 100' fallback (50g oats was being read as 5000g). Added a much smaller fallback (* 25) for generic piece-style units (portion, Stück) instead of the blanket * 100.
- GroqRecipeGeneratorService: RecipeIngredient.calories and GeneratedRecipe.calories changed from Int to Float — Groq frequently returns decimal kcal values (e.g. 247.5) which kotlinx.serialization rejects outright for an Int field, crashing recipe generation with a JSON parse error.
- Adjusted call sites (RecipeGeneratorViewModel, RecipeGeneratorScreen) to round to Int only at DB-insert/display time.

+++ app/src/main/kotlin/ch/nutrisnap/app/domain/GroqRecipeGeneratorService.kt
@@ -16,7 +16,7 @@ import java.util.concurrent.TimeUnit
 data class RecipeIngredient(
     val name: String = "",
     val amount: String = "",
-    val calories: Int = 0,
+    val calories: Float = 0f,
     val protein: Float = 0f,
     val carbs: Float = 0f,
     val fat: Float = 0f
@@ -31,7 +31,7 @@ data class GeneratedRecipe(
     val steps: List<String> = emptyList(),
     val servings: Int = 2,
     val prepTimeMinutes: Int = 30,
-    val calories: Int = 0,
+    val calories: Float = 0f,
     val protein: Float = 0f,
     val carbs: Float = 0f,
     val fat: Float = 0f
